+++ tests/integration.rs
@@ -14,9 +14,7 @@ pub fn test_rewritten_eq<D: Database>(database: &mut D, query: &str) -> bool {
     let relations = database.relations();
     let relation = Relation::try_from(parse(query).unwrap().with(&relations)).unwrap();
     let rewriten_query: &str = &ast::Query::from(&relation).to_string();
-    // DEBUG
-    relation.display_dot();
-    // Displaying the test for DEBUG purpose
+    relation.display_dot().unwrap();
     println!(
         "{}\n{}",
         format!("{query}").red(),
@@ -27,9 +25,9 @@ pub fn test_rewritten_eq<D: Database>(database: &mut D, query: &str) -> bool {
             .map(ToString::to_string)
             .join("\n")
     );
+    println!("{}", format!("{rewriten_query}").yellow());
     println!(
-        "{}\n{}",
-        format!("{rewriten_query}").yellow(),
+        "{}",
         database
             .query(rewriten_query)
             .unwrap()
@@ -62,6 +60,9 @@ const QUERIES: &[&str] = &[
     "WITH t1 AS (SELECT a,d FROM table_1),
     t2 AS (SELECT * FROM table_2)
     SELECT * FROM t1 INNER JOIN t2 ON t1.d = t2.x INNER JOIN table_2 ON t1.d=table_2.x ORDER BY t1.a, t2.x, t2.y, t2.z LIMIT 17",
+    "SELECT CASE a WHEN 5 THEN 0 ELSE a END FROM table_1",
+    "SELECT CASE WHEN a < 5 THEN 0 WHEN a < 3 THEN 3 ELSE a END FROM table_1",
+    "SELECT CASE WHEN a < 5 THEN 0 WHEN a < 3 THEN 3 END FROM table_1",
     // Test UNION
     "SELECT 1*a FROM table_1 UNION SELECT 1*x FROM table_2",
     // Test no UNION with CTEs
@@ -76,8 +77,6 @@ const QUERIES: &[&str] = &[
     "SELECT * FROM order_table LEFT JOIN item_table on id=order_id WHERE price>10",
     "SELECT UPPER(z) FROM table_2 LIMIT 5",
     "SELECT LOWER(z) FROM table_2 LIMIT 5",
-    "SELECT CHAR_LENGTH(z) AS char_length FROM table_2 LIMIT 1",
-    "SELECT POSITION('o' IN z) AS char_length FROM table_2 LIMIT 5",
 ];
 
 #[cfg(feature = "sqlite")]
@@ -102,6 +101,9 @@ const POSTGRESQL_QUERIES: &[&str] = &[
     // Test MD5
     "SELECT MD5(z) FROM table_2 LIMIT 10",
     "SELECT CONCAT(x,y,z) FROM table_2 LIMIT 11",
+    // Some joins
+    "SELECT CHAR_LENGTH(z) AS char_length FROM table_2 LIMIT 1",
+    "SELECT POSITION('o' IN z) AS char_length FROM table_2 LIMIT 5",
 ];
 
 #[test]