+++ CHANGELOG.md
@@ -6,9 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-### Changed
 ### Added
+- Conversion for `Case` expression [MR1](https://github.com/Qrlew/qrlew/pull/1)
 - Computation of the norm
 - Add `clipped_sum` transform